Update patch set 2
Patch Set 2:

From Amit: As per vendor documentation the heap minimum and maximum size should be in between 1GB to 32GB earlier it was 256MB. Also it is recommended to set minimum heap size (Xms) and maximum heap size (Xmx) to be equal to each other.

As per planned architecture for Acumos ELK, We are going keep centralized log collector on a separate server.
As we have the resource available on the server and looking at future usage I kept it 4gb. Do let me know If we want to keep it 1gb for the time Bing.
As going forward we are expecting lots and lots of logs.

Patch-set: 2
Reviewer: Gerrit User 94 <94@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=0