Update patch set 9
Patch Set 9: Code-Review+1

The docker-compose files are not specific at this point in this repo for any specific deployments, e.g. multi-node with a defined set of node assignments to services, or all-in-one. Thus the ports vs expose attributes will be incorrect for a specific deployment anyway... I will address the specific deployment customizations in the system-integration repo, where I have already done this for all-in-one and will extend that to multi-node soon. In a multi-node deploy, the compose files will be customized by the deploy script per the node assignments, initially. Then I will add docker swarm and kubernetes deployment support where this problem will be resolved finally, regardless of the node assignment of services. Thus this patch should be OK to merge as-is at this point unless there are other issues than the expose vs ports usage.

Patch-set: 9
Reviewer: Gerrit User 93 <93@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=+1