Update patch set 5
Patch Set 5:

It's frustrating that the LF CI environment is so fragile, and so complicated to get working.  We probably have to get LF help here.  You may need to open a ticket.  Or, we can cheat, and put a POM file in the top-level directory that just directs maven into the single subdirectory.

Patch-set: 5